Turn FastWrite into another WB.Write extension overload
