+++ src/System.IO.Pipelines/WritableBufferExtensions.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Runtime.CompilerServices;
+
 namespace System.IO.Pipelines
 {
     /// <summary>
@@ -43,6 +45,65 @@ namespace System.IO.Pipelines
                 }
 
                 source.Slice(offset, writable).CopyTo(buffer.Buffer.Span);
+
+                remaining -= writable;
+                offset += writable;
+
+                buffer.Advance(writable);
+            }
+        }
+
+        public static void Write(this WritableBuffer buffer, byte[] source)
+        {
+            Write(buffer, source, 0, source.Length);
+        }
+
+        public static void Write(this WritableBuffer buffer, byte[] source, int offset, int length)
+        {
+            Span<byte> dest;
+            var destLength = dest.Length;
+
+            if (destLength == 0)
+            {
+                buffer.Ensure();
+
+                // Get the new span and length
+                dest = buffer.Buffer.Span;
+                destLength = dest.Length;
+            }
+
+            var sourceLength = length;
+            if (sourceLength <= destLength)
+            {
+                ref byte pSource = ref source[offset];
+                ref byte pDest = ref dest.DangerousGetPinnableReference();
+                Unsafe.CopyBlockUnaligned(ref pDest, ref pSource, (uint)sourceLength);
+                buffer.Advance(sourceLength);
+                return;
+            }
+
+            WriteMultiBuffer(buffer, source, offset, length);
+        }
+
+        private static void WriteMultiBuffer(WritableBuffer buffer, byte[] source, int offset, int length)
+        {
+            var remaining = length;
+
+            while (remaining > 0)
+            {
+                var writable = Math.Min(remaining, buffer.Buffer.Length);
+
+                buffer.Ensure(writable);
+
+                if (writable == 0)
+                {
+                    continue;
+                }
+
+                ref byte pSource = ref source[offset];
+                ref byte pDest = ref buffer.Buffer.Span.DangerousGetPinnableReference();
+
+                Unsafe.CopyBlockUnaligned(ref pDest, ref pSource, (uint)writable);
 
                 remaining -= writable;
                 offset += writable;
