+++ tests/System.IO.Pipelines.Performance.Tests/PipeThroughput.cs
@@ -102,7 +102,7 @@ namespace System.IO.Pipelines.Performance.Tests
 
                 foreach (var write in _plaintextWrites)
                 {
-                    WriteFast(writableBuffer, write, 0, write.Length);
+                    writableBuffer.Write(write, 0, write.Length);
                 }
 
                 writableBuffer.FlushAsync().GetResult();
@@ -127,61 +127,6 @@ namespace System.IO.Pipelines.Performance.Tests
                 writableBuffer.FlushAsync().GetResult();
                 var result = _pipe.Reader.ReadAsync().GetResult();
                 _pipe.Reader.Advance(result.Buffer.End, result.Buffer.End);
-            }
-        }
-
-        [MethodImpl(MethodImplOptions.NoInlining)]
-        public static void WriteFast(WritableBuffer buffer, byte[] source, int offset, int length)
-        {
-            Span<byte> dest;
-            var destLength = dest.Length;
-
-            if (destLength == 0)
-            {
-                buffer.Ensure();
-
-                // Get the new span and length
-                dest = buffer.Buffer.Span;
-                destLength = dest.Length;
-            }
-
-            var sourceLength = length;
-            if (sourceLength <= destLength)
-            {
-                ref byte pSource = ref source[offset];
-                ref byte pDest = ref dest.DangerousGetPinnableReference();
-                Unsafe.CopyBlockUnaligned(ref pDest, ref pSource, (uint)sourceLength);
-                buffer.Advance(sourceLength);
-                return;
-            }
-
-            WriteMultiBuffer(buffer, source, offset, length);
-        }
-
-        private static void WriteMultiBuffer(WritableBuffer buffer, byte[] source, int offset, int length)
-        {
-            var remaining = length;
-
-            while (remaining > 0)
-            {
-                var writable = Math.Min(remaining, buffer.Buffer.Length);
-
-                buffer.Ensure(writable);
-
-                if (writable == 0)
-                {
-                    continue;
-                }
-
-                ref byte pSource = ref source[offset];
-                ref byte pDest = ref buffer.Buffer.Span.DangerousGetPinnableReference();
-
-                Unsafe.CopyBlockUnaligned(ref pDest, ref pSource, (uint)writable);
-
-                remaining -= writable;
-                offset += writable;
-
-                buffer.Advance(writable);
             }
         }
     }
